refactor(relay): keep rebind recovery below giant threshold
